server/requirements.txt
flask==2.3.2
Flask-cors==3.0.10
pymongo==4.3.3

# Sending emails using gmail apis
Authlib==1.2.0
emailmessage==0.2.1
python-dotenv==1.0.0
google-api-python-client==2.88.0
google-auth-httplib2==0.1.0